+++ spec/www/spec/tx-semantics-test.js
@@ -544,9 +544,12 @@ var mytests = function() {
             tx.executeSql('DROP TABLE IF EXISTS ExtraTestTable4');
             tx.executeSql('DROP TABLE IF EXISTS ExtraTestTable5');
             tx.executeSql('DROP TABLE IF EXISTS ExtraTestTable6');
+            tx.executeSql('DROP TABLE IF EXISTS AlterTestTable');
 
-            tx.executeSql('CREATE TABLE IF NOT EXISTS test_table (data)');
+            tx.executeSql('CREATE TABLE test_table (data)');
             tx.executeSql('INSERT INTO test_table VALUES (?)', ['first']);
+
+            tx.executeSql('CREATE TABLE AlterTestTable (FirstColumn)');
           }, function () {}, function () {
             db.readTransaction(function (tx) {
               tx.executeSql('SELECT * from test_table', [], function (tx, res) {
@@ -647,6 +650,21 @@ var mytests = function() {
                 function () {
                   db.readTransaction(function (tx) {
                     tx.executeSql('; ; CREATE TABLE ExtraTestTable6 (data)');
+                  }, checkDone, fail);
+                },
+                function () {
+                  db.readTransaction(function (tx) {
+                    tx.executeSql('ALTER TABLE AlterTestTable ADD COLUMN NewColumn');
+                  }, checkDone, fail);
+                },
+                function () {
+                  db.readTransaction(function (tx) {
+                    tx.executeSql('REINDEX');
+                  }, checkDone, fail);
+                },
+                function () {
+                  db.readTransaction(function (tx) {
+                    tx.executeSql('REPLACE INTO test_table VALUES ("another")');
                   }, checkDone, fail);
                 },
               ];